Create container and SVG for each manuscript page
The container is a div meant to align to the appearance of the page in
the diva viewer. The SVG is within it and provides a viewport of
dimensions equal to the page (possibly scaled). As the container moves,
the SVG can remain the same and be used as the basis of an MEI surface
element later on.